+++ test/Market.t.sol
@@ -2,11 +2,11 @@ pragma solidity ^0.8.13;
 
 import "forge-std/Test.sol";
 import "forge-std/console.sol";
-import { Market } from "../src/Market.sol";
-import { USDCtoken } from "../src/USDC.sol";
-import { MyToken } from "../src/myToken.sol";
-import { Vault } from "../src/Vault.sol";
-import { Check } from "../src/Check.sol";
+import {Market} from "../src/Market.sol";
+import {USDCtoken} from "../src/USDC.sol";
+import {MyToken} from "../src/myToken.sol";
+import {Vault} from "../src/Vault.sol";
+import {Check} from "../src/Check.sol";
 
 contract MarketTest is Test {
     Market public market;
@@ -31,7 +31,7 @@ contract MarketTest is Test {
         vm.deal(customer, 10 ether);
     }
 
-    modifier base(){
+    modifier base() {
         vm.startPrank(owner);
 
         // дали по 100 токенов каждого типа customer
@@ -46,9 +46,9 @@ contract MarketTest is Test {
 
         vm.stopPrank();
         _;
-    }    
+    }
 
-    function testMarketAddressIsntZero() public {       
+    function testMarketAddressIsntZero() public {
         assertTrue(address(market) != address(0), "adress should be not zero");
     }
 
@@ -63,16 +63,16 @@ contract MarketTest is Test {
         assertTrue(address(myToken) != address(0), "adress should be not zero");
     }
 
-    function testBalanceMoreThen0CustomerAndMarket() public base{
+    function testBalanceMoreThen0CustomerAndMarket() public base {
         assertEq(usdcToken.balanceOf(customer), 100);
         assertEq(myToken.balanceOf(customer), 100);
         assertEq(usdcToken.balanceOf(address(market)), 100);
         assertEq(myToken.balanceOf(address(market)), 100);
     }
 
-    function testBuy50Tokens() public base{
+    function testBuy50Tokens() public base {
         vm.startPrank(customer);
-        usdcToken.approve(address(market), usdcToken.balanceOf(customer));        
+        usdcToken.approve(address(market), usdcToken.balanceOf(customer));
         market.buyToken(50, address(usdcToken));
         vm.stopPrank();
 
@@ -83,19 +83,19 @@ contract MarketTest is Test {
         assertEq(usdcToken.balanceOf(address(vault)), 5);
     }
 
-    function testCrashifTokenIsntAllowed() public base{
+    function testCrashifTokenIsntAllowed() public base {
         vm.prank(owner);
         market.blockToken(address(usdcToken));
         vm.startPrank(customer);
         usdcToken.approve(address(market), usdcToken.balanceOf(customer));
-        vm.expectRevert("Token is not allowed");        
+        vm.expectRevert("Token is not allowed");
         market.buyToken(50, address(usdcToken));
         vm.stopPrank();
     }
 
-    function testBuy1tokenWith2eth() public base{
+    function testBuy1tokenWith2eth() public base {
         vm.startPrank(customer);
-        uint startBalance = customer.balance;       
+        uint256 startBalance = customer.balance;
         market.buyToken{value: 2.24 ether}();
         vm.stopPrank();
         console.log(customer.balance);
